Fix wrong integer signedness
log10() returns a non-negative value here and sizeLength is only ever
used in unsigned arithmetic, but it was stored in an int and then mixed
with size_t. Recent gcc, with -Werror=sign-conversion reject the
implicit conversion. Make sizeLength unsigned and remove the unnecessary
cast to size_t.

+++ src/stacktrace/platform/bfd-stacktrace.c
@@ -224,9 +224,9 @@ static char **stacktrace_symbols (void *const *buffer, size_t size) {
   // strings contains:
   // One big string at end containing all symbol strings.
   // `size` pointers on the big string.
-  const int sizeLength = (int)log10((double)XCP_MAX(1u, size - 1)) + 2;
+  const unsigned sizeLength = (unsigned)log10((double)XCP_MAX(1u, size - 1)) + 2;
   // `sizeLength + 1` because there is a `#` in the front of frame. ;)
-  charCount += (size_t)(sizeLength + 1) * size;
+  charCount += (sizeLength + 1) * size;
 
   char **strings = malloc(charCount + size * sizeof(char *));
   if (!strings)
